+++ typeConvert.go
@@ -21,6 +21,7 @@ package zorm
 import (
 	"context"
 	"errors"
+	"reflect"
 	"strconv"
 
 	"gitee.com/chunanyong/zorm/decimal"
@@ -128,18 +129,66 @@ func OverrideFunc(funcName string, funcObject interface{}) (bool, interface{}, e
 			oldFunc = updateEntityMap
 			updateEntityMap = newFunc
 		}
-	case "selectCount": //全局替换总条数查询
+	case "selectCount": //查询总条数
 		newFunc, ok := funcObject.(func(ctx context.Context, finder *Finder) (int, error))
 		if ok {
 			oldFunc = selectCount
 			selectCount = newFunc
 		}
-	case "wrapPageSQL": //全局替换分页语句实现
+	case "wrapPageSQL": //分页SQL
 		newFunc, ok := funcObject.(func(dialect string, sqlstr *string, page *Page) error)
 		if ok {
 			oldFunc = wrapPageSQL
 			wrapPageSQL = newFunc
 		}
+
+	case "wrapInsertSQL": //Insert IEntityStruct SQL
+		newFunc, ok := funcObject.(func(ctx context.Context, typeOf *reflect.Type, entity IEntityStruct, columns *[]reflect.StructField, values *[]interface{}) (*string, int, string, error))
+		if ok {
+			oldFunc = wrapInsertSQL
+			wrapInsertSQL = newFunc
+		}
+	case "wrapAutoIncrementInsertSQL": //IEntityStruct 主键自增值的SQL
+		newFunc, ok := funcObject.(func(pkColumnName string, sqlstr *string, dialect string, values *[]interface{}) (*int64, *int64))
+		if ok {
+			oldFunc = wrapAutoIncrementInsertSQL
+			wrapAutoIncrementInsertSQL = newFunc
+		}
+
+	case "wrapInsertSliceSQL": //批量插入 IEntityStruct 的SQL
+		newFunc, ok := funcObject.(func(ctx context.Context, config *DataSourceConfig, typeOf *reflect.Type, entityStructSlice []IEntityStruct, columns *[]reflect.StructField, values *[]interface{}) (*string, int, error))
+		if ok {
+			oldFunc = wrapInsertSliceSQL
+			wrapInsertSliceSQL = newFunc
+		}
+	case "wrapInsertEntityMapSQL": //插入 IEntityMap 的SQL
+		newFunc, ok := funcObject.(func(entity IEntityMap) (string, *[]interface{}, bool, error))
+		if ok {
+			oldFunc = wrapInsertEntityMapSQL
+			wrapInsertEntityMapSQL = newFunc
+		}
+
+	case "wrapInsertEntityMapSliceSQL": //批量插入 IEntityMap 的SQL
+		newFunc, ok := funcObject.(func(ctx context.Context, config *DataSourceConfig, entityMapSlice []IEntityMap) (*string, *[]interface{}, error))
+		if ok {
+			oldFunc = wrapInsertEntityMapSliceSQL
+			wrapInsertEntityMapSliceSQL = newFunc
+		}
+
+	case "wrapDeleteSQL": //删除 IEntityStruct 的SQL
+		newFunc, ok := funcObject.(func(entity IEntityStruct) (string, error))
+		if ok {
+			oldFunc = wrapDeleteSQL
+			wrapDeleteSQL = newFunc
+		}
+
+	case "wrapUpdateSQL": //更新 IEntityStruct 的SQL
+		newFunc, ok := funcObject.(func(typeOf *reflect.Type, entity IEntityStruct, columns *[]reflect.StructField, values *[]interface{}) (string, error))
+		if ok {
+			oldFunc = wrapUpdateSQL
+			wrapUpdateSQL = newFunc
+		}
+
 	default:
 		return false, oldFunc, errors.New("->OverrideFunc-->函数" + funcName + "暂不支持重写或不存在")
 	}