+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 v1.7.4
  - 增加 GetContextDataSourceConfig 方法,获取ctx绑定的数据源配置
+ - 感谢@CDS的反馈,修复Config取值错误
  - 完善文档,注释
 
 v1.7.3